로그인/로그아웃 기능 프로세스 변경

+++ src/main/java/com/innovation/mygym/member/application/service/AuthService.java
@@ -1,17 +1,16 @@
 package com.innovation.mygym.member.application.service;
 
+import com.innovation.mygym.exception.exception.auth.ExpiredRefreshTokenException;
+import com.innovation.mygym.member.application.dto.TokenDto;
 import com.innovation.mygym.member.application.security.AccountContext;
 import com.innovation.mygym.member.application.security.JwtAuthProvider;
 import com.innovation.mygym.member.domain.entity.RefreshToken;
-import com.innovation.mygym.member.domain.repository.MemberRepository;
 import com.innovation.mygym.member.domain.repository.RefreshTokenRepository;
-import com.innovation.mygym.member.domain.vo.Username;
 import com.innovation.mygym.member.presentation.dto.SignInRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,36 +21,52 @@ public class AuthService {
     private final AuthenticationManager authenticationManager;
     private final JwtAuthProvider jwtAuthProvider;
     private final RefreshTokenRepository refreshTokenRepository;
-    private final MemberRepository memberRepository;
 
     @Transactional
-    public String signIn(SignInRequestDto signInRequestDto) {
+    public TokenDto signIn(SignInRequestDto signInRequestDto) {
 
-        Authentication authentication = authenticationManager.authenticate(
+        Authentication authentication = getAuthentication(signInRequestDto);
+
+        Long memberId = getMemberId(authentication);
+        if (isRefreshTokenExist(memberId)) {
+            refreshTokenRepository.deleteById(memberId);
+        }
+
+        String accessToken = jwtAuthProvider.generateAccessToken(authentication);
+        String refreshToken = jwtAuthProvider.generateRefreshToken();
+
+        refreshTokenRepository.save(new RefreshToken(memberId, refreshToken));
+
+        return new TokenDto(accessToken, refreshToken);
+    }
+
+    @Transactional
+    public void signOut(String authorizationHeader) {
+
+        String accessToken = authorizationHeader.substring(7);
+
+        Long memberId = jwtAuthProvider.getUserId(accessToken);
+        if (!isRefreshTokenExist(memberId)) {
+            throw new ExpiredRefreshTokenException();
+        }
+        refreshTokenRepository.deleteById(memberId);
+    }
+
+    private Authentication getAuthentication(SignInRequestDto signInRequestDto) {
+        return authenticationManager.authenticate(
                 new UsernamePasswordAuthenticationToken(
                         signInRequestDto.username(),
                         signInRequestDto.password()
                 )
         );
-        String accessToken = jwtAuthProvider.generateAccessToken(authentication);
-        String refreshToken = jwtAuthProvider.generateRefreshToken();
+    }
 
+    private Long getMemberId(Authentication authentication) {
         AccountContext principal = (AccountContext) authentication.getPrincipal();
-        Long memberId = principal.getMemberId();
-
-        refreshTokenRepository.save(new RefreshToken(memberId, refreshToken));
-
-        return accessToken;
+        return principal.getMemberId();
     }
 
-    @Transactional
-    public void signOut(Authentication authentication) {
-
-        String username = (String) authentication.getPrincipal();
-
-        Long memberId = memberRepository.getMemberId(Username.of(username))
-                .orElseThrow(() -> new UsernameNotFoundException("사용자 정보를 다시 확인해주세요."));
-        refreshTokenRepository.deleteById(memberId);
+    private boolean isRefreshTokenExist(Long memberId) {
+        return refreshTokenRepository.existsById(memberId);
     }
-    //TODO refresh token 을 이용한 access token 갱신 구현
 }
